Move worker/worker to worker/function
