+++ worker/function/golden_dust_by_ncg.py
