--- worker/worker/common
@@ -1,1 +1,0 @@
-../../common
